Refactor and clean up code across multiple files
- Removed unnecessary includes of "signal_types.hpp" in various strategy headers.
- Changed inline constexpr to constexpr for constants in monotonic_arb.hpp.
- Updated logging functions in app.cpp to improve clarity and maintainability.
- Simplified logic in order status assignment in order_store.cpp.
- Enhanced readability by replacing iterators with more descriptive variable names in event_store.cpp.
- Added constexpr for magic numbers and improved comments for clarity in various files.
- Refactored error handling and response parsing in async_rest_connection.cpp and kalshi_rest_adapter.cpp.
- Improved function complexity handling with NOLINT comments where necessary.

+++ cpp/include/predex/oms/oms_types.hpp
@@ -123,7 +123,7 @@ enum class Outcome : std::uint8_t {
     kNo = 2,
 };
 
-enum class IntentRejectReason : std::uint16_t {
+enum class IntentRejectReason : std::uint8_t {
     kHardHalt = 0,
     kSoftHalt = 1,
     kGlobalRiskExceeded = 2,
